testexec: kill the ugly integration test wrapper
The `test-integration` target in the Makefile was an absolute disgrace.

Stop trying to be clever with shell scripts inside Make. Doing a `cd ..`
followed by a complex pipeline of `grep` and `cut` just to parse the
architecture from `limactl` output is fragile garbage. It breaks if the
output format changes even slightly, and it's impossible to debug.

Just document the standard `go test` command. We don't need a wrapper
that hides what is actually happening. If you can't type a `go test`
command line, you shouldn't be running the tests.

While I'm at it, I added `vm-rebuild`. I'm tired of running delete and
create manually every time I need a clean slate. Also added `zfs-test`
to the provision list because you obviously need the test suite to run
tests.